Time: 11 ms (83.82%), Space: 5.9 MB (44.58%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -4,7 +4,7 @@ public:
         if(x < 0)
             return false;
         int temp = x;
-        long long int revVal = 0;
+        long int revVal = 0;
         while(temp != 0)
         {
             revVal = revVal*10 + temp%10;
